+++ models/match1.cpp
@@ -1,8 +1,7 @@
 #include "match1.hpp"
 //////////////////////////// matchModel ///////////////////////////
 extern U8 level;
-inline int min(int a, int b) {return a<b?a:b;}
-inline int max(int a, int b) {return a<b?b:a;}
+
 
   void matchModel1::Update() {
         delta = false;
