+++ models/sparsematch.cpp
@@ -2,8 +2,6 @@
 
 extern U8 level;
 
-inline int min(int a, int b) {return a<b?a:b;}
-inline int max(int a, int b) {return a<b?b:a;}
 
 void SparseMatchModel::Update() {
     // update sparse hashes
