+++ paq8pxd.cpp
@@ -73,10 +73,10 @@
 
 
 // min, max functions
-#if  !defined(WINDOWS) || !defined (min)
+/*#if  !defined(WINDOWS) || !defined (min)
 inline int min(int a, int b) {return a<b?a:b;}
 inline int max(int a, int b) {return a<b?b:a;}
-#endif
+#endif*/
 
 #if defined(WINDOWS) || defined(_MSC_VER)
     #define atoll(S) _atoi64(S)
@@ -141,6 +141,11 @@ inline int max(int a, int b) {return a<b?b:a;}
 #include "models/match1.hpp"
 #include "models/match2.hpp"
 #include "models/normal.hpp"
+#include "models/chart.hpp"
+#include "models/xml.hpp"
+#include "models/nested.hpp"
+#include "models/audio.hpp"
+#include "models/linear.hpp"
 
 /////////////////////// Global context /////////////////////////
 U8 level=DEFAULT_OPTION;  // Compression level 0 to 15
@@ -3040,44 +3045,7 @@ public:
 
 
 
-class linearPredictionModel: public Model {
-  BlockData& x;
-  Buf& buf;
-   int nOLS, nLnrPrd;
-SmallStationaryContextMap sMap[3+2]{ {11,1},{11,1},{11,1},{11,1},{11,1} };
-  OLS<double, U8> ols[3]{ {32, 4, 0.995}, {32, 4, 0.995}, {32, 4, 0.995} };
-  U8 prd[5]{ 0 };
-public:
-  linearPredictionModel(BlockData& bd,U32 val=0):x(bd),buf(bd.buf),nOLS(3),nLnrPrd(3+2)  {
-  }
-  int inputs() {return nLnrPrd*2;}
-  int nets() {return 0;}
-  int netcount() {return 0;}
-  int p(Mixer& m,int val1=0,int val2=0){
-  if (x.bpos==0) {
-    const U8 W=buf(1), WW=buf(2), WWW=buf(3);
-    int i=0;
-    for (; i<nOLS; i++)
-      ols[i].Update(W);
-    for (i=1; i<=32; i++) {
-      ols[0].Add(buf(i));
-      ols[1].Add(buf(i*2-1));
-      ols[2].Add(buf(i*2));
-    }
-    for (i=0; i<nOLS; i++)
-      prd[i]=Clip(floor(ols[i].Predict()));
-    prd[i++]=Clip(W*2-WW);
-    prd[i  ]=Clip(W*3-WW*3+WWW);
-  }
-  const U8 B=x.c0<<(8-x.bpos);
-  for (int i=0; i<nLnrPrd; i++) {
-    sMap[i].set((prd[i]-B)*8+x.bpos);
-    sMap[i].mix(m, 6, 1, 2);
-  }
-  return 0;
-}
-virtual ~linearPredictionModel(){ }
-};
+
 //////////////////////////// sparseModel ///////////////////////
 
 // Model order 1-2 contexts with gaps.
@@ -4387,871 +4355,6 @@ int p(Mixer& m,int w=0,int val2=0)  {
 };
 
 
-//////////////////////////// wavModel /////////////////////////////////
-
-///////////////// Least Mean Squares predictor /////////////////
-
-template <typename F, typename T>
-class LMS {
-private:
-  F *weights, *eg, *buffer;
-  F rates[2];
-  F rho, complement, eps, prediction;
-  int S, D;
-public:
-  LMS(const int S, const int D, const F lRate, const F rRate, const F rho = 0.95, const F eps = 1e-3) : rates{ lRate, rRate }, rho(rho), complement(1. - rho), eps(eps), prediction(0.), S(S), D(D) {
-    assert(S>0 && D>0);
-    weights = new F[S+D], eg = new F[S+D], buffer = new F[S+D];
-    Reset();
-  }
-  ~LMS() {
-    delete weights, delete eg, delete buffer;
-  }
-  F Predict(const T sample)
-  {
-    memmove(&buffer[S+1], &buffer[S], (D-1) * sizeof(F));
-    buffer[S] = sample;
-    prediction = 0.;
-    for (int i=0; i<S+D; i++)
-      prediction+= weights[i] * buffer[i];
-    return prediction;
-  }
-  inline float rsqrt(const float x) {
-     float r = _mm_cvtss_f32(_mm_rsqrt_ss(_mm_set_ss(x)));
-     return (0.5f * (r + 1.0f/(x * r)));
-  }
-  void Update(const T sample)
-  {
-    const F error = sample - prediction;
-    int i=0;
-    for (; i<S; i++) {
-      const F gradient = error * buffer[i];
-      eg[i] = rho * eg[i] + complement * (gradient * gradient);
-      weights[i]+= (rates[0] * gradient * rsqrt(eg[i] + eps));
-    }
-    for (; i<S+D; i++) {
-      const F gradient = error * buffer[i];
-      eg[i] = rho * eg[i] + complement * (gradient * gradient);
-      weights[i]+= (rates[1] * gradient * rsqrt(eg[i] + eps));
-    }
-    memmove(&buffer[1], &buffer[0], (S-1) * sizeof(F));
-    buffer[0] = sample;
-  }
-  void Reset() {
-    for (int i=0; i<S+D; i++)
-      weights[i] = eg[i] = buffer[i] = 0.;
-  }
-};
-//////////////////////////// wavModel /////////////////////////////////
-
-inline U32 SQR(U32 x) {
-  return x*x;
-}
-class wavModel1: public Model {
-  BlockData& x;
-  Buf& buf;
-  private:
-  class audio8{
-BlockData& x;
-   Buf& buf;
-   const int nOLS, nLMS, nLnrPrd;
-   //nLnrPrd=14
-   SmallStationaryContextMap sMap1b[14][3]{
-    {{11,1}, {11,1}, {11,1}}, {{11,1}, {11,1}, {11,1}}, {{11,1}, {11,1}, {11,1}}, {{11,1}, {11,1}, {11,1}},
-    {{11,1}, {11,1}, {11,1}}, {{11,1}, {11,1}, {11,1}}, {{11,1}, {11,1}, {11,1}}, {{11,1}, {11,1}, {11,1}},
-    {{11,1}, {11,1}, {11,1}}, {{11,1}, {11,1}, {11,1}}, {{11,1}, {11,1}, {11,1}}, {{11,1}, {11,1}, {11,1}},
-    {{11,1}, {11,1}, {11,1}}, {{11,1}, {11,1}, {11,1}}
-  };
-  //nOLS=8
-  OLS<double, int8_t> ols[8][2]{
-    {{128, 24, 0.9975}, {128, 24, 0.9975}},
-    {{90, 30, 0.9965}, {90, 30, 0.9965}},
-    {{90, 31, 0.996}, {90, 31, 0.996}},
-    {{90, 32, 0.995}, {90, 32, 0.995}},
-    {{90, 33, 0.995}, {90, 33, 0.995}},
-    {{90, 34, 0.9985}, {90, 34, 0.9985}},
-    {{28, 4, 0.98}, {28, 4, 0.98}},
-    {{28, 3, 0.992}, {28, 3, 0.992}}
-  };
-  //nLMS=3
-  LMS<float, int8_t> lms[3][2]{
-    {{1280, 640, 3e-5, 2e-5}, {1280, 640, 3e-5, 2e-5}},
-    {{640, 64, 8e-5, 1e-5}, {640, 64, 8e-5, 1e-5}},
-    {{2450, 8, 1.6e-5, 1e-6}, {2450, 8, 1.6e-5, 1e-6}}
-  };
-  //nLnrPrd=14
-  int prd[14][2][2];
-  int residuals[14][2];
-    int stereo, ch;
-    U32 mask, errLog, mxCtx;
-    int S;
-  int wmode;
-    public:
-   audio8(BlockData& bd):x(bd),buf(bd.buf),nOLS(8), nLMS(3), nLnrPrd(nOLS+nLMS+3),prd{ 0 },residuals{ 0 },stereo(0), ch(0),mask(0), errLog(0), mxCtx(0),S(0),wmode(0){
-      
-    }   
-    
-inline int s2(int i) { return int(short(buf(i)+256*buf(i-1))); }
-inline int t2(int i) { return int(short(buf(i-1)+256*buf(i))); }
-
-inline int X1(int i) {
-  switch (wmode) {
-    case 0: return buf(i)-128;
-    case 1: return buf(i<<1)-128;
-    case 2: return s2(i<<1);
-    case 3: return s2(i<<2);
-    case 4: return (buf(i)^128)-128;
-    case 5: return (buf(i<<1)^128)-128;
-    case 6: return t2(i<<1);
-    case 7: return t2(i<<2);
-    default: return 0;
-  }
-}
-
-inline int X2(int i) {
-  switch (wmode) {
-    case 0: return buf(i+S)-128;
-    case 1: return buf((i<<1)-1)-128;
-    case 2: return s2((i+S)<<1);
-    case 3: return s2((i<<2)-2);
-    case 4: return (buf(i+S)^128)-128;
-    case 5: return (buf((i<<1)-1)^128)-128;
-    case 6: return t2((i+S)<<1);
-    case 7: return t2((i<<2)-2);
-    default: return 0;
-  }
-}
-
-inline int signedClip8(const int i) {
-  return max(-128, min(127, i));
-}
-
-void audio8bModel(Mixer& m, int info) {
-  const int8_t B = x.c0<<(8-x.bpos);
-   if (x.blpos==0 && x.bpos==1) {
-  //if (x.bpos==1 &&x.blpos==0) {
-      assert((info&2)==0);
-      stereo = (info&1);
-      mask = 0;
-      wmode=info;
-      for (int i=0; i<nLMS; i++)
-        lms[i][0].Reset(), lms[i][1].Reset();
-    }
-  if (x.bpos==0) {
-    ch=(stereo)?x.blpos&1:0;
-    const int8_t s = int(((wmode&4)>0)?buf(1)^128:buf(1))-128;
-    const int pCh = ch^stereo;
-    int i = 0;
-    for (errLog=0; i<nOLS; i++) {
-      ols[i][pCh].Update(s);
-      residuals[i][pCh] = s-prd[i][pCh][0];
-      const U32 absResidual = (U32)abs(residuals[i][pCh]);
-      mask+=mask+(absResidual>4);
-      errLog+=SQR(absResidual);
-    }
-    for (int j=0; j<nLMS; j++)
-      lms[j][pCh].Update(s);
-    for (; i<nLnrPrd; i++)
-      residuals[i][pCh] = s-prd[i][pCh][0];
-    errLog = min(0xF, ilog2(errLog));
-    mxCtx = ilog2(min(0x1F, BitCount(mask)))*2+ch;
-
-    int k1=90, k2=k1-12*stereo;
-    for (int j=(i=1); j<=k1; j++, i+=1<<((j>8)+(j>16)+(j>64))) ols[1][ch].Add(X1(i));
-    for (int j=(i=1); j<=k2; j++, i+=1<<((j>5)+(j>10)+(j>17)+(j>26)+(j>37))) ols[2][ch].Add(X1(i));
-    for (int j=(i=1); j<=k2; j++, i+=1<<((j>3)+(j>7)+(j>14)+(j>20)+(j>33)+(j>49))) ols[3][ch].Add(X1(i));
-    for (int j=(i=1); j<=k2; j++, i+=1+(j>4)+(j>8)) ols[4][ch].Add(X1(i));
-    for (int j=(i=1); j<=k1; j++, i+=2+((j>3)+(j>9)+(j>19)+(j>36)+(j>61))) ols[5][ch].Add(X1(i));
-    if (stereo) {
-      for (i=1; i<=k1-k2; i++) {
-        const double s = (double)X2(i);
-        ols[2][ch].AddFloat(s);
-        ols[3][ch].AddFloat(s);
-        ols[4][ch].AddFloat(s);
-      }
-    }
-    k1=28, k2=k1-6*stereo;
-    for (i=1; i<=k2; i++) {
-      const double s = (double)X1(i);
-      ols[0][ch].AddFloat(s);
-      ols[6][ch].AddFloat(s);
-      ols[7][ch].AddFloat(s);
-    }
-    for (; i<=96; i++) ols[0][ch].Add(X1(i));
-    if (stereo) {
-      for (i=1; i<=k1-k2; i++) {
-        const double s = (double)X2(i);
-        ols[0][ch].AddFloat(s);
-        ols[6][ch].AddFloat(s);
-        ols[7][ch].AddFloat(s);
-      }
-      for (; i<=32; i++) ols[0][ch].Add(X2(i));
-    }
-    else
-      for (; i<=128; i++) ols[0][ch].Add(X1(i));
-
-    for (i=0; i<nOLS; i++)
-      prd[i][ch][0] = signedClip8((int)floor(ols[i][ch].Predict()));
-    for (; i<nOLS+nLMS; i++)
-      prd[i][ch][0] = signedClip8((int)floor(lms[i-nOLS][ch].Predict(s)));
-    prd[i++][ch][0] = signedClip8(X1(1)*2-X1(2));
-    prd[i++][ch][0] = signedClip8(X1(1)*3-X1(2)*3+X1(3));
-    prd[i  ][ch][0] = signedClip8(X1(1)*4-X1(2)*6+X1(3)*4-X1(4));
-    for (i=0; i<nLnrPrd; i++)
-      prd[i][ch][1] = signedClip8(prd[i][ch][0]+residuals[i][pCh]);
-  }
-  for (int i=0; i<nLnrPrd; i++) {
-    const U32 ctx = (prd[i][ch][0]-B)*8+x.bpos;
-    sMap1b[i][0].set(ctx);
-    sMap1b[i][1].set(ctx);
-    sMap1b[i][2].set((prd[i][ch][1]-B)*8+x.bpos);
-    sMap1b[i][0].mix(m, 6, 1, 2+(i>=nOLS));
-    sMap1b[i][1].mix(m, 9, 1, 2+(i>=nOLS));
-    sMap1b[i][2].mix(m, 7, 1, 3);
-  }
-  m.set((errLog<<8)|x.c0, 4096);
-  m.set((U8(mask)<<3)|(ch<<2)|(x.bpos>>1), 2048);
-  m.set((mxCtx<<7)|(buf(1)>>1), 1280);
-  m.set((errLog<<4)|(ch<<3)|x.bpos, 256);
-  m.set(mxCtx, 10);
-}
-};
-
-//16 bit audio model
-class audio16{
-BlockData& x;
-   Buf& buf;
-   const int nOLS, nLMS, nLnrPrd;
-   //nLnrPrd=14
-   SmallStationaryContextMap sMap1b[14][4]{
-    {{17,1},{17,1},{17,1},{17,1}}, {{17,1},{17,1},{17,1},{17,1}}, {{17,1},{17,1},{17,1},{17,1}}, {{17,1},{17,1},{17,1},{17,1}},
-    {{17,1},{17,1},{17,1},{17,1}}, {{17,1},{17,1},{17,1},{17,1}}, {{17,1},{17,1},{17,1},{17,1}}, {{17,1},{17,1},{17,1},{17,1}},
-    {{17,1},{17,1},{17,1},{17,1}}, {{17,1},{17,1},{17,1},{17,1}}, {{17,1},{17,1},{17,1},{17,1}}, {{17,1},{17,1},{17,1},{17,1}},
-    {{17,1},{17,1},{17,1},{17,1}}, {{17,1},{17,1},{17,1},{17,1}}
-  };
-  //nOLS=8
-  OLS<double, int16_t> ols[8][2]{
-    {{128, 24, 0.9975}, {128, 24, 0.9975}},
-    {{90, 30, 0.997}, {90, 30, 0.997}},
-    {{90, 31, 0.996}, {90, 31, 0.996}},
-    {{90, 32, 0.995}, {90, 32, 0.995}},
-    {{90, 33, 0.995}, {90, 33, 0.995}},
-    {{90, 34, 0.9985}, {90, 34, 0.9985}},
-    {{28, 4, 0.98}, {28, 4, 0.98}},
-    {{32, 3, 0.992}, {32, 3, 0.992}}
-  };
-  //nLMS=3
-  LMS<float, int16_t> lms[3][2]{
-    {{1280, 640, 5e-5f, 5e-5f}, {1280, 640, 5e-5f, 5e-5f}},
-    {{640, 64, 7e-5f, 1e-5f}, {640, 64, 7e-5f, 1e-5f}},
-    {{2450, 8, 2e-5f, 2e-6f}, {2450, 8, 2e-5f, 2e-6f}}
-  };
-  //nLnrPrd=14
-  int prd[14][2][2];
-  int residuals[14][2];
-    int stereo, ch,lsb;
-    U32 mask, errLog, mxCtx;
-    int S;
-  int wmode;
-  int16_t sample ;
-    public:
-   audio16(BlockData& bd):x(bd),buf(bd.buf),nOLS(8), nLMS(3), nLnrPrd(nOLS+nLMS+3),prd{ 0 },residuals{ 0 },stereo(0), ch(0),lsb(0),mask(0), errLog(0), mxCtx(0),S(0),wmode(0),sample(0){
-      for (int i=0; i<nLMS; i++)
-        lms[i][0].Reset(), lms[i][1].Reset();
-    }   
-    
-inline int s2(int i) { return int(short(buf(i)+256*buf(i-1))); }
-inline int t2(int i) { return int(short(buf(i-1)+256*buf(i))); }
-
-inline int X1(int i) {
-  switch (wmode) {
-    case 0: return buf(i)-128;
-    case 1: return buf(i<<1)-128;
-    case 2: return s2(i<<1);
-    case 3: return s2(i<<2);
-    case 4: return (buf(i)^128)-128;
-    case 5: return (buf(i<<1)^128)-128;
-    case 6: return t2(i<<1);
-    case 7: return t2(i<<2);
-    default: return 0;
-  }
-}
-
-inline int X2(int i) {
-  switch (wmode) {
-    case 0: return buf(i+S)-128;
-    case 1: return buf((i<<1)-1)-128;
-    case 2: return s2((i+S)<<1);
-    case 3: return s2((i<<2)-2);
-    case 4: return (buf(i+S)^128)-128;
-    case 5: return (buf((i<<1)-1)^128)-128;
-    case 6: return t2((i+S)<<1);
-    case 7: return t2((i<<2)-2);
-    default: return 0;
-  }
-}
-
-inline int signedClip16(const int i) {
-  return max(-32768, min(32767, i));
-}
-void audio16bModel(Mixer& m, int info) {
-  if (x.blpos==0 && x.bpos==1) {
-  //if (x.blpos==1 && x.bpos==1) {
-     //info|=4;  // comment this line if skipping the endianness transform
-     assert((info&2)!=0);
-      stereo = (info&1);
-      lsb = (info<4);
-      mask = 0;
-      wmode=info;
-      for (int i=0; i<nLMS; i++)
-        lms[i][0].Reset(), lms[i][1].Reset();
-    }
-  if (x.bpos==0) {
-      ch=(stereo)?(x.blpos&2)>>1:0;
-      lsb=(x.blpos&1)^(wmode<4);
-      if ((x.blpos&1)==0)
-       {
-        sample = (wmode<4)?s2(2):t2(2);
-        const int pCh = ch^stereo;
-        int i = 0;
-        for (errLog=0; i<nOLS; i++) {
-          ols[i][pCh].Update(sample);
-          residuals[i][pCh] = sample-prd[i][pCh][0];
-          const U32 absResidual = (U32)abs(residuals[i][pCh]);
-          mask+=mask+(absResidual>128);
-          errLog+=SQR(absResidual>>6);
-        }
-        for (int j=0; j<nLMS; j++){
-          lms[j][pCh].Update(sample); }
-        for (; i<nLnrPrd; i++){
-          residuals[i][pCh] = sample-prd[i][pCh][0]; }
-        errLog = min(0xF, ilog2(errLog));
-
-        if (stereo) {
-          for (int i=1; i<=24; i++) ols[0][ch].Add(X2(i));
-          for (int i=1; i<=104; i++) ols[0][ch].Add(X1(i));
-        }
-        else
-          for (int i=1; i<=128; i++) ols[0][ch].Add(X1(i));
-
-        int k1=90, k2=k1-12*stereo;
-        for (int j=(i=1); j<=k1; j++, i+=1<<((j>16)+(j>32)+(j>64))) ols[1][ch].Add(X1(i));
-        for (int j=(i=1); j<=k2; j++, i+=1<<((j>5)+(j>10)+(j>17)+(j>26)+(j>37))) ols[2][ch].Add(X1(i));       
-        for (int j=(i=1); j<=k2; j++, i+=1<<((j>3)+(j>7)+(j>14)+(j>20)+(j>33)+(j>49))) ols[3][ch].Add(X1(i));
-        for (int j=(i=1); j<=k2; j++, i+=1+(j>4)+(j>8)) ols[4][ch].Add(X1(i));
-        for (int j=(i=1); j<=k1; j++, i+=2+((j>3)+(j>9)+(j>19)+(j>36)+(j>61))) ols[5][ch].Add(X1(i));
-
-        if (stereo) {
-          for (i=1; i<=k1-k2; i++) {
-            const double s = (double)X2(i);
-            ols[2][ch].AddFloat(s);
-            ols[3][ch].AddFloat(s);
-            ols[4][ch].AddFloat(s);
-          }
-        }
-
-        k1=28, k2=k1-6*stereo;
-        for (i=1; i<=k2; i++) ols[6][ch].Add(X1(i));
-        for (i=1; i<=k1-k2; i++) ols[6][ch].Add(X2(i));
-
-        k1=32, k2=k1-8*stereo;
-        for (i=1; i<=k2; i++) ols[7][ch].Add(X1(i));
-        for (i=1; i<=k1-k2; i++) ols[7][ch].Add(X2(i));
-
-        for (i=0; i<nOLS; i++)
-          prd[i][ch][0] = signedClip16((int)floor(ols[i][ch].Predict()));
-        for (; i<nOLS+nLMS; i++)
-          prd[i][ch][0] = signedClip16((int)floor(lms[i-nOLS][ch].Predict(sample)));
-        prd[i++][ch][0] = signedClip16(X1(1)*2-X1(2));
-        prd[i++][ch][0] = signedClip16(X1(1)*3-X1(2)*3+X1(3));
-        prd[i  ][ch][0] = signedClip16(X1(1)*4-X1(2)*6+X1(3)*4-X1(4));
-        for (i=0; i<nLnrPrd; i++)
-          prd[i][ch][1] = signedClip16(prd[i][ch][0]+residuals[i][pCh]);
-      }
-      mxCtx = ilog2(min(0x1F, BitCount(mask)))*4+ch*2+lsb;
-  }
-
-  const int16_t B = int16_t( (info<4)? (lsb)?U8(x.c0<<(8-x.bpos)):(x.c0<<(16-x.bpos))|buf(1) : (lsb)?(buf(1)<<8)|U8(x.c0<<(8-x.bpos)):x.c0<<(16-x.bpos) );
-
-  for (int i=0; i<nLnrPrd; i++) {
-    const U32 ctx0 = U16(prd[i][ch][0]-B);
-    const U32 ctx1 = U16(prd[i][ch][1]-B);
-
-    sMap1b[i][0].set( (lsb<<16)|(x.bpos<<13)|(ctx0>>(3<<(!lsb))) );
-    sMap1b[i][1].set( (lsb<<16)|(x.bpos<<13)|(ctx0>>((!lsb)+(3<<(!lsb)))) );
-    sMap1b[i][2].set( (lsb<<16)|(x.bpos<<13)|(ctx0>>((!lsb)*2+(3<<(!lsb)))) );
-    sMap1b[i][3].set( (lsb<<16)|(x.bpos<<13)|(ctx1>>((!lsb)+(3<<(!lsb)))) );
-
-    sMap1b[i][0].mix(m, 7, 1/*, 2+(i>=nOLS)*/);
-    sMap1b[i][1].mix(m, 10, 1/*, 2+(i>=nOLS)*/);
-    sMap1b[i][2].mix(m, 6, 1/*, 3+(i>=nOLS)*/);
-    sMap1b[i][3].mix(m, 6, 1/*, 2+(i>=nOLS)*/);
-  }
-
-  m.set((errLog<<9)|(lsb<<8)|x.c0, 8192);
-  m.set((U8(mask)<<4)|(ch<<3)|(lsb<<2)|(x.bpos>>1), 4096);
-  m.set((mxCtx<<7)|(buf(1)>>1), 2560);
-  m.set((errLog<<4)|(ch<<3)|(lsb<<2)|(x.bpos>>1), 256);
-  m.set(mxCtx, 20);
-}
-};
-
-
-audio8 *a8mode;
-audio16 *a16mode;
-public:
-  wavModel1(BlockData& bd): x(bd),buf(bd.buf){
-    a8mode=new audio8(bd);
-    a16mode=new audio16(bd);
-}
-int inputs() {
-    return (14*4)*2;
-}
-int nets() {
-     return 8192+ 4096+ 2560+ 256+ 20;
-} 
-  int netcount() {return 5;}
-int p(Mixer& m,int info,int val2=0){
-    if ((info&2)==0) {
-    a8mode->audio8bModel(m, info);
-    } 
-    else  {
-    a16mode->audio16bModel(m, info);
-    }
-  return 0;
-  } 
-
-  virtual ~wavModel1(){
-      delete a8mode;
-      delete a16mode;
-   }
- 
-};
-
-
-class nestModel1: public Model {
-  BlockData& x;
-  Buf& buf;
-  int ic, bc, pc,vc, qc, lvc, wc,ac, ec, uc, sense1, sense2, w;
-  const int N;
-  ContextMap cm;
-public:
-  nestModel1(BlockData& bd,U32 val=0):x(bd),buf(bd.buf), ic(0), bc(0),
-   pc(0),vc(0), qc(0), lvc(0), wc(0),ac(0), ec(0), uc(0), sense1(0), sense2(0), w(0), N(12),
-   cm(CMlimit(level>8?0x800000 :(MEM()/2) ), N,M_NEST)  {
-  }
-  int inputs() {return N*cm.inputs();}
-  int nets() {return 512;}
-  int netcount() {return 1;}
-int p(Mixer& m,int val1=0,int val2=0){
-    if (x.filetype==DBASE ||x.filetype==HDR ||x.filetype==DECA || x.filetype==ARM|| x.filetype==IMGUNK){
-        if (val2==-1) return 1;
-        for (int i=0; i<inputs(); i++)
-        m.add(0);
-        return 0;
-    }
-    if (x.bpos==0) {
-    int c=x.c4&255, matched=1, vv;
-    w*=((vc&7)>0 && (vc&7)<3);
-    if (c&0x80) w = w*11*32 + c;
-    const int lc = (c >= 'A' && c <= 'Z'?c+'a'-'A':c) ;
-    if (lc == 'a' || lc == 'e' || lc == 'i' || lc == 'o' || lc == 'u'){ vv = 1; w = w*997*8 + (lc/4-22); } else
-    if (lc >= 'a' && lc <= 'z' || c>128){ vv = 2; w = w*271*32 + lc-97; } else
-    if (lc == ' ' || lc == '.' || lc == ',' || lc == '\n'|| lc == 5) vv = 3; else
-    if (lc >= '0' && lc <= '9') vv = 4; else
-    if (lc == 'y') vv = 5; else
-    if (lc == '\'') vv = 6; else vv=(c&32)?7:0;
-    vc = (vc << 3) | vv;
-    if (vv != lvc) {
-      wc = (wc << 3) | vv;
-      lvc = vv;
-    }
-    switch(c) {
-      case ' ': qc = 0; break;
-      case '(': ic += 31; break;
-      case ')': ic -= 31; break;
-      case '[': ic += 11; break;
-      case ']': ic -= 11; break;
-      case '<': ic += 23; qc += 34; break;
-      case '>': ic -= 23; qc /= 5; break;
-      case ':': pc = 20; break;
-      case '{': ic += 17; break;
-      case '}': ic -= 17; break;
-      case '|': pc += 223; break;
-      case '"': pc += 0x40; break;
-      case '\'': pc += 0x42; if (c!=(U8)(x.c4>>8)) sense2^=1; else ac+=(2*sense2-1); break;
-      case 5: 
-      case '\n': pc = qc = 0; break;
-      case '.': pc = 0; break;
-      case '!': pc = 0; break;
-      case '?': pc = 0; break;
-      case '#': pc += 0x08; break;
-      case '%': pc += 0x76; break;
-      case '$': pc += 0x45; break;
-      case '*': pc += 0x35; break;
-      case '-': pc += 0x3; break;
-      case '@': pc += 0x72; break;
-      case '&': qc += 0x12; break;
-      case ';': qc /= 3; break;
-      case '\\': pc += 0x29; break;
-      case '/': pc += 0x11;
-                if (buf.size() > 1 && buf(1) == '<') qc += 74;
-                break;
-      case '=': pc += 87; if (c!=(U8)(x.c4>>8)) sense1^=1; else ec+=(2*sense1-1); break;
-      default: matched = 0;
-    }
-    if (x.c4==0x266C743B) uc=min(7,uc+1);
-    else if (x.c4==0x2667743B) uc-=(uc>0);
-    if (matched) bc = 0; else bc += 1;
-    if (bc > 300) bc = ic = pc = qc = uc = 0;
-if (val2==-1) return 1;
-    cm.set(hash( (vv>0 && vv<3)?0:(lc|0x100), ic&0x3FF, ec&0x7, ac&0x7, uc ));
-    cm.set(hash(ic, w, ilog2(bc+1)));
-    
-    cm.set(U32((3*vc+77*pc+373*ic+qc)&0xffff));
-    cm.set(U32((31*vc+27*pc+281*qc)&0xffff));
-    cm.set(U32((13*vc+271*ic+qc+bc)&0xffff));
-    cm.set(U32((13*vc+ic)&0xffff));
-    cm.set(U32((vc/3+pc)&0xffff));
-    cm.set(U32((17*pc+7*ic)&0xffff));
-    cm.set(U32((7*wc+qc)&0xffff));
-    cm.set(U32((vc&0xffff)|((x.f4&0xf)<<16)));
-    cm.set(U32(((3*pc)&0xffff)|((x.f4&0xf)<<16)));
-    cm.set(U32((ic&0xffff)|((x.f4&0xf)<<16)));
-  }
-    
-    cm.mix(m);
-    m.set(vc&511,512);
-  return 0;
-}
-virtual ~nestModel1(){ }
-};
-
-
-/*
-====== XML model ======
-*/
-
-
-struct XMLAttribute {
-  U32 Name, Value, Length;
-};
-
-struct XMLContent {
-  U32 Data, Length, Type;
-};
-
-struct XMLTag {
-  U32 Name, Length;
-  int Level;
-  bool EndTag, Empty;
-  XMLContent Content;
-  struct XMLAttributes {
-    XMLAttribute Items[4];
-    U32 Index;
-  } Attributes;    
-};
-
-struct XMLTagCache {
-  XMLTag Tags[CacheSize];
-  U32 Index;
-};
-
-enum ContentFlags {
-  Text        = 0x001,
-  Number      = 0x002,
-  Date        = 0x004,
-  Time        = 0x008,
-  URL         = 0x010,
-  Link        = 0x020,
-  Coordinates = 0x040,
-  Temperature = 0x080,
-  ISBN        = 0x100
-};
-
-enum XMLState {
-  None               = 0,
-  ReadTagName        = 1,
-  ReadTag            = 2,
-  ReadAttributeName  = 3,
-  ReadAttributeValue = 4,
-  ReadContent        = 5,
-  ReadCDATA          = 6,
-  ReadComment        = 7
-};
-
-
-#define DetectContent(){ \
-  if ((x.c4&0xF0F0F0F0)==0x30303030){ \
-    int i = 0, j = 0; \
-    while ((i<4) && ( (j=(x.c4>>(8*i))&0xFF)>=0x30 && j<=0x39 )) \
-      i++; \
-\
-    if (i==4 && ( ((c8&0xFDF0F0FD)==0x2D30302D && buf(9)>=0x30 && buf(9)<=0x39) || ((c8&0xF0FDF0FD)==0x302D302D) )) \
-      (*Content).Type |= Date; \
-  } \
-  else if (((c8&0xF0F0FDF0)==0x30302D30 || (c8&0xF0F0F0FD)==0x3030302D) && buf(9)>=0x30 && buf(9)<=0x39){ \
-    int i = 2, j = 0; \
-    while ((i<4) && ( (j=(c8>>(8*i))&0xFF)>=0x30 && j<=0x39 )) \
-      i++; \
-\
-    if (i==4 && (x.c4&0xF0FDF0F0)==0x302D3030) \
-      (*Content).Type |= Date; \
-  } \
-\
-  if ((x.c4&0xF0FFF0F0)==0x303A3030 && buf(5)>=0x30 && buf(5)<=0x39 && ((buf(6)<0x30 || buf(6)>0x39) || ((c8&0xF0F0FF00)==0x30303A00 && (buf(9)<0x30 || buf(9)>0x39)))) \
-    (*Content).Type |= Time; \
-\
-  if ((*Content).Length>=8 && (c8&0x80808080)==0 && (x.c4&0x80808080)==0) \
-    (*Content).Type |= Text; \
-\
-  if ((c8&0xF0F0FF)==0x3030C2 && (x.c4&0xFFF0F0FF)==0xB0303027){ \
-    int i = 2; \
-    while ((i<7) && buf(i)>=0x30 && buf(i)<=0x39) \
-      i+=(i&1)*2+1; \
-\
-    if (i==10) \
-      (*Content).Type |= Coordinates; \
-  } \
-\
-  if ((x.c4&0xFFFFFA)==0xC2B042 && B!=0x47 && (((x.c4>>24)>=0x30 && (x.c4>>24)<=0x39) || ((x.c4>>24)==0x20 && (buf(5)>=0x30 && buf(5)<=0x39)))) \
-    (*Content).Type |= Temperature; \
-\
-  if (B>=0x30 && B<=0x39) \
-    (*Content).Type |= Number; \
-\
-  if (x.c4==0x4953424E && buf(5)==0x20) \
-    (*Content).Type |= ISBN; \
-} 
-class XMLModel1: public Model {
-  BlockData& x;
-  Buf& buf;
-  XMLTagCache Cache;
-  XMLState State, pState;
-  U32 c8, WhiteSpaceRun, pWSRun, IndentTab, IndentStep, LineEnding,lastState;
-  ContextMap cm;
-  U32 StateBH[8],stateContext;
-  XMLTag *pTag,*Tag;
-    XMLAttribute *Attribute;
-    XMLContent *Content ;
-public:
-  XMLModel1(BlockData& bd,U32 val=0):x(bd),buf(bd.buf), State(None), pState(None), c8(0),
-   WhiteSpaceRun(0), pWSRun(0), IndentTab(0), IndentStep(2), LineEnding(2),lastState(0), cm(level>9?0x1000000:CMlimit(MEM()/4), 4,M_XML) {
-       memset(&Cache, 0, sizeof(XMLTagCache));
-       memset(&StateBH, 0, sizeof(StateBH));  
-        
-  }
-  int inputs() {return 4*cm.inputs();}
-  int nets() {return 0;}
-  int netcount() {return 0;}
-  void setContexts(){
-      stateContext=0;
-      U8 B = (U8)x.c4;
-       pTag = &Cache.Tags[ (Cache.Index-1)&(CacheSize-1) ], Tag = &Cache.Tags[ Cache.Index&(CacheSize-1) ];
-     Attribute = &((*Tag).Attributes.Items[ (*Tag).Attributes.Index&3 ]);
-     Content = &(*Tag).Content;     
-    //XMLTag *pTag = &Cache.Tags[ (Cache.Index-1)&(CacheSize-1) ], *Tag = &Cache.Tags[ Cache.Index&(CacheSize-1) ];
-    //XMLAttribute *Attribute = &((*Tag).Attributes.Items[ (*Tag).Attributes.Index&3 ]);
-    //XMLContent *Content = &(*Tag).Content;
-    pState = State;
-    c8 = (c8<<8)|buf(5);
-    if ((B==0x09 || B==0x20) && (B==(U8)(x.c4>>8) || !WhiteSpaceRun)){
-      WhiteSpaceRun++;
-      IndentTab = (B==0x09);
-    }
-    else{
-      if ((State==None || (State==ReadContent && (*Content).Length<=LineEnding+WhiteSpaceRun)) && WhiteSpaceRun>1+IndentTab && WhiteSpaceRun!=pWSRun){
-        IndentStep=abs((int)(WhiteSpaceRun-pWSRun));
-        pWSRun = WhiteSpaceRun;
-      }
-      WhiteSpaceRun=0;
-    }
-    if (B==0x0A ||(B==5 && ((U8)(x.c4>>8)!='&')) )
-      LineEnding = 1+((U8)(x.c4>>8)==0x0D);
-    if(State!=None) lastState=buf.pos;
-    // if &< or &> or &</ then ignore and force to reading content, so wordmodel knows
-    if (  (x.c4&&0xffff)==0x263c || (x.c4&&0xffff)==0x263e|| (x.c4&&0xffffff)==0x263c2f){
-              State = ReadContent;
-    }
-    switch (State){
-      case None : {
-        if (B==0x3C){
-          State = ReadTagName;
-          memset(Tag, 0, sizeof(XMLTag));
-          (*Tag).Level = ((*pTag).EndTag || (*pTag).Empty)?(*pTag).Level:(*pTag).Level+1;
-        }
-        if ((*Tag).Level>1)
-          DetectContent();
-        
-        stateContext=(hash(pState, State, ((*pTag).Level+1)*IndentStep - WhiteSpaceRun));
-        break;
-      }
-      case ReadTagName : {
-        if ((*Tag).Length>0 && (B==0x09 || B==0x0A || B==0x0D || (B==5 && ((U8)(x.c4>>8)!='&')) || B==0x20))
-          State = ReadTag;
-        else if ((B>127)|| (B==0x3A || (B>='A' && B<='Z') || B==0x5F|| /*B==1|| B==2 ||*/ (B>='a' && B<='z')) || ((*Tag).Length>0 && (B==0x2D || B==0x2E || (B>='0' && B<='9')))){
-          (*Tag).Length++;
-          (*Tag).Name = (*Tag).Name * 263 * 32 + (B&0xDF);
-        }
-        else if (B == 0x3E){
-          if ((*Tag).EndTag){
-            State = None;
-            Cache.Index++;
-          }
-          else
-            State = ReadContent;
-        }
-        else if (B!=0x21 && B!=0x2D && B!=0x2F && B!=0x5B){
-          State = None;
-          Cache.Index++;
-        }
-        else if ((*Tag).Length==0){
-          if (B==0x2F){
-            (*Tag).EndTag = true;
-            (*Tag).Level = max(0,(*Tag).Level-1);
-          }
-          else if (x.c4==0x3C212D2D){
-            State = ReadComment;
-            (*Tag).Level = max(0,(*Tag).Level-1);
-          }
-        }
-
-        if ((*Tag).Length==1 && (x.c4&0xFFFF00)==0x3C2100){
-          memset(Tag, 0, sizeof(XMLTag));
-          State = None;
-        }
-        else if ((*Tag).Length==5 && c8==0x215B4344 && x.c4==0x4154415B){
-          State = ReadCDATA;
-          (*Tag).Level = max(0,(*Tag).Level-1);
-        }
-        
-        int i = 1;
-        do{
-          pTag = &Cache.Tags[ (Cache.Index-i)&(CacheSize-1) ];
-          i+=1+((*pTag).EndTag && Cache.Tags[ (Cache.Index-i-1)&(CacheSize-1) ].Name==(*pTag).Name);
-        }
-        while ( i<CacheSize && ((*pTag).EndTag || (*pTag).Empty) );
-
-        stateContext=(hash(pState*8+State, (*Tag).Name, (*Tag).Level, (*pTag).Name, (*pTag).Level!=(*Tag).Level ));
-        break;
-      }
-      case ReadTag : {
-        if (B==0x2F)
-          (*Tag).Empty = true;
-        else if (B==0x3E){
-          if ((*Tag).Empty){
-            State = None;
-            Cache.Index++;
-          }
-          else
-            State = ReadContent;
-        }
-        else if (B!=0x09 && B!=0x0A && B!=0x0D && B!=0x20 && B!=5){
-          State = ReadAttributeName;
-          (*Attribute).Name = B&0xDF;
-        }
-        stateContext=(hash(pState, State, (*Tag).Name, B, (*Tag).Attributes.Index ));
-        break;
-      }
-      case ReadAttributeName : {
-        if ((x.c4&0xFFF0)==0x3D20 && (B==0x22 || B==0x27)){
-          State = ReadAttributeValue;
-          if ((c8&0xDFDF)==0x4852 && (x.c4&0xDFDF0000)==0x45460000)
-            (*Content).Type |= Link;
-        }
-        else if (B!=0x22 && B!=0x27 && B!=0x3D)
-          (*Attribute).Name = (*Attribute).Name * 263 * 32 + (B&0xDF);
-
-        stateContext=(hash(pState*8+State, (*Attribute).Name, (*Tag).Attributes.Index, (*Tag).Name, (*Content).Type ));
-        break;
-      }
-      case ReadAttributeValue : {
-        if (B==0x22 || B==0x27){
-          (*Tag).Attributes.Index++;
-          State = ReadTag;
-        }
-        else{
-          (*Attribute).Value = (*Attribute).Value* 263 * 32 + (B&0xDF);
-          (*Attribute).Length++;
-          if ((c8&0xDFDFDFDF)==0x48545450 && ((x.c4>>8)==0x3A2F2F || x.c4==0x733A2F2F)) // HTTP :// s://
-            (*Content).Type |= URL;
-        }
-        stateContext=(hash(pState, State, (*Attribute).Name, (*Content).Type ));
-        break;
-      }
-      case ReadContent : {
-        if (B==0x3C && ((x.c4&0xffff)!=0x263c)){ // if new tag and not escape
-          State = ReadTagName;
-          Cache.Index++;
-          memset(&Cache.Tags[ Cache.Index&(CacheSize-1) ], 0, sizeof(XMLTag));
-          Cache.Tags[ Cache.Index&(CacheSize-1) ].Level = (*Tag).Level+1;
-        }
-        else{
-          (*Content).Length++;
-          (*Content).Data = (*Content).Data * 997*16 + (B&0xDF);
-
-          DetectContent();
-        }
-        stateContext=(hash(pState, State, (*Tag).Name, x.c4&0xC0FF ));
-        break;
-      }
-      case ReadCDATA : {
-        if ((x.c4&0xFFFFFF)==0x5D5D3E){ // ]]>
-          State = None;
-          Cache.Index++;
-        }
-        stateContext=(hash(pState, State));
-        break;
-      }
-      case ReadComment : {
-        if ((x.c4&0xFFFFFF)==0x2D2D3E){ // -->
-          State = None;
-          Cache.Index++;
-        }
-        stateContext=(hash(pState, State));
-        break;
-      }
-    }
-    StateBH[pState] = (StateBH[pState]<<8)|B;
-    pTag = &Cache.Tags[ (Cache.Index-1)&(CacheSize-1) ];
-  }
-int p(Mixer& m,int val1=0,int val2=0){
-    if (x.filetype==DBASE ||x.filetype==HDR ||x.filetype==DECA || x.filetype==ARM  || x.filetype==IMGUNK|| x.filetype==BINTEXT){
-        if (val2==-1) return 1;
-        for (int i=0; i<inputs(); ++i)
-        m.add(0);
-        return 0;
-    }
-    
-    if (x.bpos==0) {
-    setContexts();
-    if (val2==-1) return 1;
-     pTag = &Cache.Tags[ (Cache.Index-1)&(CacheSize-1) ], Tag = &Cache.Tags[ Cache.Index&(CacheSize-1) ];
-     //*Attribute = &((*Tag).Attributes.Items[ (*Tag).Attributes.Index&3 ]);
-    // *Content = &(*Tag).Content;     
-    cm.set(stateContext);
-    // set context if last state was less then 256 bytes ago
-    if ((buf.pos-lastState)<256){ 
-        cm.set(hash(State, (*Tag).Level, pState*2+(*Tag).EndTag, (*Tag).Name));
-        cm.set(hash((*pTag).Name, State*2+(*pTag).EndTag, (*pTag).Content.Type, (*Tag).Content.Type));
-        cm.set(hash(State*2+(*Tag).EndTag, (*Tag).Name, (*Tag).Content.Type, x.c4&0xE0FF));
-    }else {
-        cm.set();
-        cm.set();
-        cm.set();
-    } 
-  }
-   if (val2==0)  cm.mix(m);
-
-  U8 s = ((StateBH[State]>>(28-x.bpos))&0x08) |
-         ((StateBH[State]>>(21-x.bpos))&0x04) |
-         ((StateBH[State]>>(14-x.bpos))&0x02) |
-         ((StateBH[State]>>( 7-x.bpos))&0x01) |
-         ((x.bpos)<<4);
-  return (s<<3)|State;
-}
-virtual ~XMLModel1(){ }
-};
 
 
 
@@ -5285,108 +4388,7 @@ int p(Mixer& m,int val1=0,int val2=0){
 
 
 
-///////////////////////////// chartModel ////////////////////////////
-class chartModel: public Model {
-  BlockData& x;
-  Buf& buf;
-  ContextMap cm,cn;
-  Array<U32> chart;
-  Array<U8> indirect;
-  Array<U8> indirect2;
-  Array<U8> indirect3;
-public:
-  chartModel(BlockData& bd,U32 val=0):x(bd),buf(bd.buf),cm(CMlimit(MEM()*4),30,M_CHART),cn(CMlimit(MEM()),20,M_CHART), 
-  chart(32), indirect(2048), indirect2(256), indirect3(0x10000) 
- {   
- }
- int inputs() {return 30*cm.inputs()+20*cn.inputs();}
-int nets() {return 0;}
-  int netcount() {return 0;}
-int p(Mixer& m,int val1=0,int val2=0){
-   if (!x.bpos){
-       U32 c4=x.c4;
-    const
-      U32 w = c4&65535,
-      w0 = c4&16777215,
-      w1 = c4&255,
-      w2 = c4<<8&65280,
-      w3 = c4<<8&16711680,
-      w4 = c4<<8&4278190080,
-      a[3]={c4>>5&460551,    c4>>2&460551,   c4&197379},
-      b[3]={c4>>23&448|      c4>>18&56|      c4>>13&7,
-            c4>>20&448|      c4>>15&56|      c4>>10&7,
-            c4>>18&48 |      c4>>12&12|      c4>>8&3},
-      d[3]={c4>>15&448|      c4>>10&56|      c4>>5&7,
-            c4>>12&448|      c4>>7&56 |      c4>>2&7,
-            c4>>10&48 |      c4>>4&12 |      c4&3},
-      c[3]={c4&255,          c4>>8&255,      c4>>16&255};
-U32 p=0;
-    for (
-      int i=0,
-      j=0,
-      f=b[0],
-      e=a[0];
 
-    i<3;
-    ++i
-      
-    ){
-         j=(i<<9);
-      f=j|b[i];
-      
-      e=a[i];
-      indirect[f&0x7FF]=w1; // <--Update indirect
-      const int g =indirect[(j|d[i])&0x7FF];
-      chart[(i<<3|e>>16&255)&31]=w0; // <--Fix chart
-      chart[(i<<3|e>>8&255)&31]=w<<8; // <--Update chart
-      cn.set(U32(e&7|(((e>>8)&7)<<3)|(((e>>16)&7)<<6)));// low 3 bits of last 3 bytes
-      //Maps[p++].set(g);
-      //cn.set(e); // <--Model previous/current/next slot
-      //cn.set(g); // <--Guesses next "c4&0xFF"
-      cn.set(U32(w2|g)); // <--Guesses next "c4&0xFFFF"
-      cn.set(U32(w3|g)); // <--Guesses next "c4&0xFF00FF"
-      cn.set(U32(w4|g)); // <--Guesses next "c4&0xFF0000FF"
-      //cm.set(c[i]); // <--Models buf(1,2,3)
-    }
-
-    indirect2[buf(2)]=buf(1);
-    int g=indirect2[buf(1)];
-    cn.set(U32(g)); // <--Guesses next "c4&0xFF"
-    cn.set(U32(w2|g)); // <--Guesses next "c4&0xFFFF"
-    cn.set(U32(w3|g)); // <--Guesses next "c4&0xFF00FF"
-    cn.set(U32(w4|g)); // <--Guesses next "c4&0xFF0000FF"
-
-    indirect3[buf(3)<<8|buf(2)]=buf(1);
-    g=indirect3[buf(2)<<8|buf(1)];
-    cn.set(U32(g)); // <--Guesses next "c4&0xFF"
-    cn.set(U32(w2|g)); // <--Guesses next "c4&0xFFFF"
-    cn.set(U32(w3|g)); // <--Guesses next "c4&0xFF00FF"
-    cn.set(U32(w4|g)); // <--Guesses next "c4&0xFF0000FF"
-
-
-    for (
-      int
-      i=10,
-      s=0,
-      e=a[0],
-      k=chart[0];
-
-    i<20;
-    s=++i>>3,
-      e=a[s],
-      k=chart[i]
-    ){ // k e
-      cm.set(U32(k<<s)); // 111 000
-      cm.set(hash(e,k,s)); // 111 111
-      cm.set((hash(e&255,k>>16)^(k&255))<<s); // 101 001
-    }
-  }
-  cn.mix(m);
-  cm.mix(m);
-  return 0;
-}
-  virtual ~chartModel(){ }
-};
 
 // LSTM
 #include "lstm1.inc"
